chore: fix semantic issues, move edit and feature-details to needs refactoring folder
