+++ src/needs-refactoring/modules/edit/hooks/useNextAccessibleToilet.ts
