+++ src/needs-refactoring/modules/feature-panel/components/FeatureDescription.tsx
