+++ src/needs-refactoring/modules/feature-panel/components/HeaderImageSection.tsx
