+++ src/needs-refactoring/pages/[placeType]/[id]/report/wheelchair-accessibility.tsx
@@ -1,7 +1,7 @@
 import React, { useContext } from "react";
 import { getLayout } from "~/layouts/DefaultLayout";
-import { WheelchairEditorOld } from "~/modules/edit/components/WheelchairEditorOld";
 import { FeaturePanelContext } from "~/needs-refactoring/components/CombinedFeaturePanel/FeaturePanelContext";
+import { WheelchairEditorOld } from "~/needs-refactoring/modules/edit/components/WheelchairEditorOld";
 
 function WheelchairAccessibility() {
   const { features } = useContext(FeaturePanelContext);
